Modify upgrade for mo_snapshot
