+++ pkg/bootstrap/versions/v1_2_2/tenant_upgrade_list.go
@@ -1,0 +1,45 @@
+// Copyright 2024 Matrix Origin
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package v1_2_2
+
+import (
+	"fmt"
+	"github.com/matrixorigin/matrixone/pkg/bootstrap/versions"
+	"github.com/matrixorigin/matrixone/pkg/catalog"
+	"github.com/matrixorigin/matrixone/pkg/util/executor"
+)
+
+var tenantUpgEntries = []versions.UpgradeEntry{
+	upg_mo_snapshots,
+}
+
+var upg_mo_snapshots = versions.UpgradeEntry{
+	Schema:    catalog.MO_CATALOG,
+	TableName: catalog.MO_SNAPSHOTS,
+	UpgType:   versions.MODIFY_COLUMN,
+	PreSql:    fmt.Sprintf("alter table %s.%s drop column ts;", catalog.MO_CATALOG, catalog.MO_SNAPSHOTS),
+	UpgSql:    fmt.Sprintf("alter table %s.%s add column ts bigint unsigned after sname", catalog.MO_CATALOG, catalog.MO_SNAPSHOTS),
+	CheckFunc: func(txn executor.TxnExecutor, accountId uint32) (bool, error) {
+		colInfo, err := versions.CheckTableColumn(txn, accountId, catalog.MO_CATALOG, catalog.MO_SNAPSHOTS, "ts")
+		if err != nil {
+			return false, err
+		}
+
+		if colInfo.ColType != "TIMESTAMP" {
+			return true, nil
+		}
+		return false, nil
+	},
+}
