+++ pkg/catalog/types.go
@@ -136,8 +136,6 @@ const (
 
 	// MO_SNAPSHOTS
 	MO_SNAPSHOTS = "mo_snapshots"
-
-	MO_SNAPSHOTS_TS = "ts"
 )
 
 const (
